Feature/159

+++ platform-api/src/main/java/com/flow/platform/api/service/job/JobService.java
@@ -50,7 +50,7 @@ public interface JobService {
      *
      * @return job with children node result
      */
-    Job find(String path, Integer number);
+    Job find(String path, Long number);
 
     /**
      * Find by job id
@@ -67,7 +67,7 @@ public interface JobService {
     /**
      * Get job yml content
      */
-    String findYml(String path, Integer number);
+    String findYml(String path, Long number);
 
     /**
      * delete jobs by flowPath
@@ -115,7 +115,7 @@ public interface JobService {
     /**
      * stop job
      */
-    Job stop(String name, Integer buildNumber);
+    Job stop(String name, Long buildNumber);
 
     /**
      * update job
